+++ src/Map/MapMarker.tsx
@@ -1,6 +1,6 @@
 import React, { Fragment, useRef, useState, FC } from 'react';
 import { Tooltip } from 'realayers';
-import { motion } from 'framer-motion';
+import { motion } from 'framer-motion/dist/framer-motion';
 import css from './MapMarker.module.css';
 
 export interface MapMarkerProps {