+++ src/TreeMap/TreeMapRect.tsx
@@ -1,12 +1,11 @@
 import React, { FC, ReactElement, useRef, useState, Fragment } from 'react';
-import { motion } from 'framer-motion';
+import { motion } from 'framer-motion/dist/framer-motion';
 import chroma from 'chroma-js';
 import { ChartTooltip, ChartTooltipProps } from '../common/Tooltip';
 import { CloneElement } from 'rdk';
 import { DEFAULT_TRANSITION } from '../common/Motion';
 
 export interface TreeMapRectProps {
-
   /**
    * The internal data object built by d3
    */
@@ -40,17 +39,17 @@ export interface TreeMapRectProps {
   /**
    * Event for when the arc is clicked.
    */
-   onClick?: (event, data) => void;
+  onClick?: (event, data) => void;
 
-   /**
-    * Event for when the arc has mouse enter.
-    */
-   onMouseEnter?: (event, data) => void;
+  /**
+   * Event for when the arc has mouse enter.
+   */
+  onMouseEnter?: (event, data) => void;
 
-   /**
-    * Event for when the arc has mouse leave.
-    */
-   onMouseLeave?: (event, data) => void;
+  /**
+   * Event for when the arc has mouse leave.
+   */
+  onMouseLeave?: (event, data) => void;
 }
 
 export const TreeMapRect: FC<Partial<TreeMapRectProps>> = ({
@@ -66,9 +65,7 @@ export const TreeMapRect: FC<Partial<TreeMapRectProps>> = ({
   const [internalActive, setInternalActive] = useState<boolean>(false);
   const rectRef = useRef<any | null>(null);
   const transition = animated ? DEFAULT_TRANSITION : { type: false, delay: 0 };
-  const currentFill = internalActive
-    ? chroma(fill).darken(0.8).hex()
-    : fill;
+  const currentFill = internalActive ? chroma(fill).darken(0.8).hex() : fill;
 
   return (
     <Fragment>
@@ -86,14 +83,14 @@ export const TreeMapRect: FC<Partial<TreeMapRectProps>> = ({
         }}
         style={{ cursor }}
         transition={transition}
-        onClick={event => {
+        onClick={(event) => {
           onClick?.(event, data);
         }}
-        onMouseEnter={event => {
+        onMouseEnter={(event) => {
           setInternalActive(true);
           onMouseEnter?.(event, data);
         }}
-        onMouseLeave={event => {
+        onMouseLeave={(event) => {
           setInternalActive(false);
           onMouseLeave?.(event, data);
         }}